client/web: signal need to wait for auth across tabs
This implements a signaling mechanism that indicates the need to
call the auth/session/wait endpoint when returning to the app (in
any tab). We try to open the auth URL in a new tab so that
auth/session/wait can be called once the current tab regains focus.
However, it might open in the same tab based on browser settings,
thus foregoing the subsequent call to auth/session/wait.

Fixes #11905

Signed-off-by: Gesa Stupperich <gesa@tailscale.com>