clientupdate: support updating to release candidates
Adds a new track for release candidates which is mapped to a new
Updater Arguments field: acceptReleaseCandidates.

When calling update, if the "release-candidate" track is provided,
both the stable and release-candidate tracks are checked for updates.
The newer of the two versions is selected.

When calling version with the --upstream and --accept-release-candidates
flags, the latest release-candidate version is shown if it is newer
than the stable version.

Alpine updates to release candidates are not yet supported.

updates #18193